refactor(utils): file rename
change convertUnixEpoch to convertUnixEpochToDate
